split out long running scaling test
